# backend/src/main/resources/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Seax 工作流程與組織記憶 API",
    "version": "0.2.0",
    "description": "GlobalMemory schemaVersion 2：版本化知識、結案案例、scope、證據、UNKNOWN 原因及回饋診斷。POC 沿用 /api/v1；既有外部消費者需同步升級。必填文字不得只有空白。所有 POST 須 Idempotency-Key。Job、交易、重試及結案規則沿用 backend/spec.md；AI_INPUT_TOO_LARGE 不可重試，不以 UNKNOWN 掩蓋技術失敗。"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/global-memory/knowledge-items": {
      "get": {
        "summary": "查詢有來源及範圍的組織知識",
        "operationId": "getKnowledgeItemPage_global-memory_knowledge-items",
        "tags": [
          "global-memory"
        ],
        "responses": {
          "200": {
            "description": "查詢有來源及範圍的組織知識",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItemPage"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": true
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "required": false
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "required": false
          },
          {
            "name": "departmentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "FEATURE",
                "RESPONSIBILITY",
                "CAPABILITY",
                "COMMON_RULE",
                "PROJECT_ARRANGEMENT"
              ]
            },
            "required": false
          },
          {
            "name": "scopeLevel",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ORGANIZATION",
                "PROJECT"
              ]
            },
            "required": false
          },
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false
          }
        ],
        "description": "必須指定固定 version，多頁沿用同一版本；limit 預設 50、範圍 1–200，offset 非負。舊版本成員與內容不隨新版本變動。"
      }
    },
    "/global-memory/relationships": {
      "get": {
        "summary": "查詢有方向與來源的部門關係",
        "operationId": "getRelationshipPage_global-memory_relationships",
        "tags": [
          "global-memory"
        ],
        "responses": {
          "200": {
            "description": "查詢有方向與來源的部門關係",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationshipPage"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": true
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "required": false
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "required": false
          },
          {
            "name": "departmentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "REPORTS_TO",
                "UPSTREAM_OF",
                "COLLABORATES_WITH"
              ]
            },
            "required": false
          }
        ],
        "description": "必須指定固定 version，多頁沿用同一版本；limit 預設 50、範圍 1–200，offset 非負。舊版本成員與內容不隨新版本變動。"
      }
    },
    "/global-memory/experiences": {
      "get": {
        "summary": "查詢已納入指定記憶版本的結案案例摘要",
        "operationId": "getExperiencePage_global-memory_experiences",
        "tags": [
          "global-memory"
        ],
        "responses": {
          "200": {
            "description": "查詢已納入指定記憶版本的結案案例摘要",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperiencePage"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": true
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "required": false
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "required": false
          },
          {
            "name": "departmentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false
          },
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false
          }
        ],
        "description": "必須指定固定 version，多頁沿用同一版本；limit 預設 50、範圍 1–200，offset 非負。舊版本成員與內容不隨新版本變動。 案例依 closedAt、projectId 降冪；部門篩選只計 ASSIGNED。"
      }
    },
    "/global-memory/experiences/{projectId}": {
      "get": {
        "summary": "查詢指定版本的完整結案案例",
        "operationId": "getExperienceResult_global-memory_experiences_projectId",
        "tags": [
          "global-memory"
        ],
        "responses": {
          "200": {
            "description": "查詢指定版本的完整結案案例",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperienceResult"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": true
          },
          {
            "name": "projectId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "description": "不在該版本的案例回傳 404；REPORT_APPROVED 不宣稱工作已執行。"
      }
    },
    "/global-memory/evidence": {
      "get": {
        "summary": "查詢可核對的原始來源與摘錄",
        "operationId": "getEvidencePage_global-memory_evidence",
        "tags": [
          "global-memory"
        ],
        "responses": {
          "200": {
            "description": "查詢可核對的原始來源與摘錄",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidencePage"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": true
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "required": false
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "required": false
          },
          {
            "name": "knowledgeItemId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false
          },
          {
            "name": "relationshipId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false
          },
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false
          }
        ],
        "description": "必須指定固定 version，多頁沿用同一版本；limit 預設 50、範圍 1–200，offset 非負。舊版本成員與內容不隨新版本變動。"
      }
    },
...
    "/projects/{projectId}/analysis-results": {
      "get": {
        "summary": "查詢不可變 AI 歸屬判斷、資訊缺口及引用",
        "operationId": "getAnalysisRecordPage_projects_projectId_analysis-results",
        "tags": [
          "projects"
        ],
        "responses": {
          "200": {
            "description": "查詢不可變 AI 歸屬判斷、資訊缺口及引用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRecordPage"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "required": false
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "required": false
          },
          {
            "name": "jobId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false
          },
          {
            "name": "projectId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "description": "依 reportVersion 升冪。人工改派不修改歷史判斷；過往 explanation 不代表目前人工歸屬理由。"
      }
    },
...
    "/jobs/{jobId}/feedback-result": {
      "get": {
        "summary": "查詢成功回饋發布結果及未採納診斷",
        "operationId": "getFeedbackResult_jobs_jobId_feedback-result",
        "tags": [
          "jobs"
        ],
        "responses": {
          "200": {
            "description": "查詢成功回饋發布結果及未採納診斷",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResult"
                }
              }
            }
          },
          "400": {
            "description": "參數或 JSON 不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "資源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "狀態、版本、相依或冪等衝突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "JSON body 超過 10 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "未預期錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "description": "FEEDBACK SUCCEEDED 才回傳 200。未成功回 409 FEEDBACK_RESULT_NOT_READY；非 FEEDBACK 回 409 WRONG_JOB_TYPE。不回傳供應商原始錯誤。"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "KnowledgeItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "FEATURE",
              "RESPONSIBILITY",
              "CAPABILITY",
              "COMMON_RULE",
              "PROJECT_ARRANGEMENT"
            ]
          },
          "statement": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000
          },
          "departmentIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "uniqueItems": true
          },
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "evidenceIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "minItems": 1,
            "uniqueItems": true
          }
        },
        "required": [
          "id",
          "type",
          "statement",
          "departmentIds",
          "scope",
          "evidenceIds"
        ],
        "additionalProperties": false
      },
      "DepartmentRelationship": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "REPORTS_TO",
              "UPSTREAM_OF",
              "COLLABORATES_WITH"
            ]
          },
          "fromDepartmentId": {
            "type": "string",
            "format": "uuid"
          },
          "toDepartmentId": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000
          },
          "exchangedItems": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2000
            },
            "maxItems": 50
          },
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "evidenceIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "minItems": 1,
            "uniqueItems": true
          }
        },
        "required": [
          "id",
          "type",
          "fromDepartmentId",
          "toDepartmentId",
          "description",
          "exchangedItems",
          "scope",
          "evidenceIds"
        ],
        "additionalProperties": false
      },
...
      "ExperienceSummary": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "projectName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "reportId": {
            "type": "string",
            "format": "uuid"
          },
          "reportVersion": {
            "type": "integer",
            "minimum": 0
          },
          "closedAt": {
            "type": "string",
            "format": "date-time"
          },
          "experienceMeaning": {
            "type": "string",
            "const": "REPORT_APPROVED"
          },
          "workflowCount": {
            "type": "integer",
            "minimum": 0
          },
          "departmentIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "projectId",
          "projectName",
          "reportId",
          "reportVersion",
          "closedAt",
          "experienceMeaning",
          "workflowCount",
          "departmentIds"
        ],
        "additionalProperties": false
      },
      "ExperienceResult": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "experience": {
            "$ref": "#/components/schemas/ProjectExperience"
          }
        },
        "required": [
          "version",
          "experience"
        ],
        "additionalProperties": false
      },
...
      "AnalysisRecordPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnalysisRecord"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "additionalProperties": false
      },
...
      "FeedbackResult": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "globalMemoryVersion": {
            "type": "integer",
            "minimum": 1
          },
          "publication": {
            "$ref": "#/components/schemas/FeedbackPublication"
          },
          "diagnostics": {
            "$ref": "#/components/schemas/FeedbackDiagnostics"
          }
        },
        "required": [
          "jobId",
          "globalMemoryVersion",
          "publication",
          "diagnostics"
        ],
        "additionalProperties": false
      },
      "KnowledgeItemPage": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KnowledgeItem"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "version",
          "items",
          "total",
          "limit",
          "offset"
        ],
        "additionalProperties": false
      },
      "RelationshipPage": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DepartmentRelationship"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "version",
          "items",
          "total",
          "limit",
          "offset"
        ],
        "additionalProperties": false
      },
      "ExperiencePage": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExperienceSummary"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "version",
          "items",
          "total",
          "limit",
          "offset"
        ],
        "additionalProperties": false
      },
      "EvidencePage": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Evidence"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "version",
          "items",
          "total",
          "limit",
          "offset"
        ],
        "additionalProperties": false
      }
    }
  }
}